requirements.txt
Django==2.2.13
gunicorn==19.9.0
mysql-connector==2.2.9
mysqlclient==1.4.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
djangorestframework
django-stubs
mypy